Feature Add: Ajout d'un paquet de pages dont j'aurai besoin à long terme. Je vais fermer ce pull request et ouvrir un autre issue pour commencer à générer ces pages
