+++ client/src/App.tsx
@@ -4,7 +4,8 @@ import Histoire from './pages/Histoire';
 import VisualisationInvenataire from './pages/VisualisationInventaire';
 import Reglements from './pages/Reglements';
 import EnsemblesReglements from './pages/EnsemblesReglements';
-
+import EnsRegTerritoire from './pages/EnsRegTerr';
+import CompReg from './pages/CompReg';
 const app: React.FC = () => {
   return (
       <Router>
@@ -14,6 +15,8 @@ const app: React.FC = () => {
           <Route path="/inventaire" element={<VisualisationInvenataire/>}/>
           <Route path="/reg" element={<Reglements/>}/>
           <Route path="/ens-reg" element={<EnsemblesReglements/>}/>
+          <Route path="/ens-reg-terr" element={<EnsRegTerritoire/>}/>
+          <Route path="/comp-reg" element={<CompReg/>}/>
         </Routes>
       </Router>
   );
