+++ client/src/pages/CompReg.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import MenuBar from '../components/MenuBar';
+
+
+const CompReg:React.FC = () =>{
+    return(
+        <div className="page-comp-reg">
+            <MenuBar/>
+        </div>
+    )
+}
+export default CompReg;
