+++ client/src/pages/EnsRegTerr.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import MenuBar from '../components/MenuBar';
+
+
+const EnsRegTerritoire: React.FC =()=>{
+    return(
+        <div className="page-ens-reg-terr">
+            <MenuBar/>
+        </div>
+    )
+}
+
+export default EnsRegTerritoire;
