+++ client/src/pages/common.css
@@ -63,6 +63,8 @@ body,html{
 /* ---------------------------------------------------------------------------------------- */
 .ville-control{
     padding: 10px 20px;
+    flex-direction: row;
+    display:flex;
 }
